Add files via upload
Dyslexia is a neurological disorder that is characterized by imprecise comprehension of words and generally poor reading performance. It affects a significant population of school-age children, putting them at risk of poor academic performance and low self-esteem for a lifetime. Dyslexia is characterized by difficulties with reading, poor spelling and decoding abilities, writing challenges, Difficulty with Phonological Processing, Difficulty with Sequencing and Organization etc...
The overarching goal project is to develop a dyslexia diagnostic method informed by neural biomarkers. Early detection and intervention are critical to mitigating the impact of dyslexia on individuals' lives.

 Project aims to develop a diagnostic method for dyslexia using machine learning and deep learning techniques. Also,We must address challenges like finding interpretable biomarkers, data
privacy, and overfitting.Goal is to achieve clinically relevant dyslexia detection for improved outcomes.